refactor: refactor path for DTOs to use all folders
