+++ CarsApi.API/Controllers/BrandController.cs
@@ -2,7 +2,9 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using CarsApi.Application.DTOs;
+using CarsApi.Application.DTOs.Create;
+using CarsApi.Application.DTOs.Response;
+using CarsApi.Application.DTOs.Update;
 using CarsApi.Domain.Services;
 using Microsoft.AspNetCore.Mvc;
