+++ CarsApi.Application/DTOs/Create/CreateBrandDto.cs
@@ -1,9 +1,4 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace CarsApi.Application.DTOs
+namespace CarsApi.Application.DTOs.Create
 {
     public record BrandCreateDto(string Name);
 }
