+++ CarsApi.Application/DTOs/Create/CreateModelDto.cs
@@ -1,7 +1,2 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
 namespace CarsApi.Application.DTOs.Create
 {   public record ModelCreateDto(int Brand_id,string Name,float Fipe_value); }
